+++ essencium-backend/src/main/java/de/frachtwerk/essencium/backend/controller/RoleController.java
@@ -20,14 +20,15 @@
 package de.frachtwerk.essencium.backend.controller;
 
 import de.frachtwerk.essencium.backend.model.Role;
-import de.frachtwerk.essencium.backend.model.dto.RoleDto;
 import de.frachtwerk.essencium.backend.model.exception.DuplicateResourceException;
+import de.frachtwerk.essencium.backend.model.exception.ResourceUpdateException;
 import de.frachtwerk.essencium.backend.service.RoleService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Set;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -62,44 +63,47 @@ public class RoleController {
   @PreAuthorize("hasPermission(#id, 'Role', 'read')")
   @Operation(description = "Retrieve a specific role by its id")
   public Role findById(@PathVariable("id") @NotNull final String id) {
-    return roleService.getById(id);
+    return roleService.getByName(id);
   }
 
   @PostMapping
   @PreAuthorize("hasPermission(#role, 'create')")
   @ResponseStatus(HttpStatus.CREATED)
   @Operation(description = "Create a new role")
-  public Role create(@Valid @RequestBody @NotNull final RoleDto role) {
-    if (roleService.getRole(role.getName()).isPresent()) {
+  public Role create(@Valid @RequestBody @NotNull final Role role) {
+    if (Objects.nonNull(roleService.getByName(role.getName()))) {
       throw new DuplicateResourceException("already existing");
     }
-    return roleService.create(role);
+    return roleService.save(role);
   }
 
-  @PutMapping(value = "/{id}")
-  @PreAuthorize("hasPermission(#id, 'Role', 'update')")
+  @PutMapping(value = "/{name}")
+  @PreAuthorize("hasPermission(#name, 'Role', 'update')")
   @Operation(description = "Update a given role by passing an entire update object")
   public Role updateObject(
-      @PathVariable("id") @NotNull final String id,
-      @Valid @RequestBody @NotNull final RoleDto role) {
-    return roleService.update(id, role);
+      @PathVariable("name") @NotNull final String name,
+      @Valid @RequestBody @NotNull final Role role) {
+    if (!role.getName().equals(name)) {
+      throw new ResourceUpdateException("Name needs to match entity name");
+    }
+    return roleService.save(role);
   }
 
-  @PatchMapping(value = "/{id}")
-  @PreAuthorize("hasPermission(#id, 'Role', 'update')")
+  @PatchMapping(value = "/{name}")
+  @PreAuthorize("hasPermission(#name, 'Role', 'update')")
   @Operation(description = "Update a given role by passing individual fields")
   public Role update(
-      @PathVariable("id") final String id,
+      @PathVariable("name") final String name,
       @NotNull @RequestBody final Map<String, Object> roleFields) {
-    return roleService.patch(id, roleFields);
+    return roleService.patch(name, roleFields);
   }
 
-  @DeleteMapping(value = "/{id}")
-  @PreAuthorize("hasPermission(#id, 'Role', 'delete')")
+  @DeleteMapping(value = "/{name}")
+  @PreAuthorize("hasPermission(#name, 'Role', 'delete')")
   @ResponseStatus(HttpStatus.NO_CONTENT)
   @Operation(description = "Delete a given role by its id")
-  public void delete(@PathVariable("id") @NotNull final String id) {
-    roleService.deleteById(id);
+  public void delete(@PathVariable("name") @NotNull final String name) {
+    roleService.deleteById(name);
   }
 
   @RequestMapping(value = "/**", method = RequestMethod.OPTIONS)