+++ essencium-backend/src/test/java/de/frachtwerk/essencium/backend/service/UUIDUserService.java
@@ -24,7 +24,9 @@ import de.frachtwerk.essencium.backend.model.TestUUIDUser;
 import de.frachtwerk.essencium.backend.model.dto.UserDto;
 import de.frachtwerk.essencium.backend.repository.BaseUserRepository;
 import jakarta.validation.constraints.NotNull;
+import java.util.HashSet;
 import java.util.UUID;
+import java.util.stream.Collectors;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
 public class UUIDUserService extends AbstractUserService<TestUUIDUser, UUID, UserDto<UUID>> {
@@ -40,12 +42,14 @@ public class UUIDUserService extends AbstractUserService<TestUUIDUser, UUID, Use
 
   @Override
   protected @NotNull <E extends UserDto<UUID>> TestUUIDUser convertDtoToEntity(@NotNull E entity) {
-
-    Role role = roleService.getById(entity.getRole());
+    HashSet<Role> roles =
+        entity.getRoles().stream()
+            .map(roleService::getByName)
+            .collect(Collectors.toCollection(HashSet::new));
     return TestUUIDUser.builder()
         .email(entity.getEmail())
         .enabled(entity.isEnabled())
-        .role(role)
+        .roles(roles)
         .firstName(entity.getFirstName())
         .lastName(entity.getLastName())
         .locale(entity.getLocale())