+++ essencium-backend/src/testIntegration/java/de/frachtwerk/essencium/backend/test/integration/security/UserSecurityTest.java
@@ -25,7 +25,6 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import de.frachtwerk.essencium.backend.model.Role;
 import de.frachtwerk.essencium.backend.model.dto.LoginRequest;
 import de.frachtwerk.essencium.backend.repository.RightRepository;
 import de.frachtwerk.essencium.backend.repository.RoleRepository;
@@ -33,7 +32,6 @@ import de.frachtwerk.essencium.backend.test.integration.IntegrationTestApplicati
 import de.frachtwerk.essencium.backend.test.integration.model.TestUser;
 import de.frachtwerk.essencium.backend.test.integration.repository.TestBaseUserRepository;
 import de.frachtwerk.essencium.backend.test.integration.util.TestingUtils;
-import java.util.HashSet;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -65,19 +63,11 @@ class UserSecurityTest {
 
   @BeforeEach
   public void setupSingle() {
+    testingUtils.createOrGetAdminRole();
     if (testUser == null) {
       testUser = testingUtils.getOrCreateAdminUser();
     }
-    testUser.setRole(
-        roleRepository
-            .findByName("ADMIN")
-            .orElse(
-                roleRepository.save(
-                    Role.builder()
-                        .name("ADMIN")
-                        .description("Application Admin")
-                        .rights(new HashSet<>(rightRepository.findAll()))
-                        .build())));
+    testUser.getRoles().add(roleRepository.findByName("ADMIN"));
     testUser = userRepository.save(testUser);
   }